core_ai_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
tiktoken
SQLAlchemy